New DTX that works in all modes (SILK/CELT/HYBRID)
A frame is marked as valid for DTX if it contains noise or only digital
silence. As before, there is an overhang period of 200 ms and a maximum
consecutive DTX period of 400 ms.

If the new DTX cannot be used because of the complexity setting and
sampling frequency chosen, the SILK DTX will be used instead.

Signed-off-by: Jean-Marc Valin <jmvalin@jmvalin.ca>