gitlab-ci: Add doc job to invoke the build_draft scripts.
All the drafts are currently published, so this is not
particularly important. However, the code bitrots if not
tested regularly, so this will remind us to maintain what
we have in the tree.

xml2rfc is installed with pip.

Using a python base image makes for a smaller download than
installing python on the gcc image used for the other build
jobs.